Adjust rollback comment, move rollback funcs to rollback.go
